requirements.txt
Django==2.2.7
djangorestframework==3.10.3
djangorestframework-camel-case==1.1.2
django-cors-headers==3.2.0